refactor(api): split auto-label routes into focused modules
